requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
flask-cors==6.0.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.4
flask-mysqldb
PyJWT==2.10.1
reportlab==4.0.9
python-dotenv==1.0.0
pymupdf
Flask-Mail
